+++ android/src/main/java/ly/count/android/sdk/react/CountlyMessagingService.java
@@ -8,6 +8,7 @@ import android.util.Log;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
+import ly.count.android.sdk.messaging.CountlyConfigPush;
 import ly.count.android.sdk.messaging.CountlyPush;
 import ly.count.android.sdk.Countly;
 
@@ -32,17 +33,13 @@ public class CountlyMessagingService extends FirebaseMessagingService {
         }
 
         if(!Countly.sharedInstance().isInitialized()) {
-            int mode = CountlyPush.getLastMessagingMethod(this);
             Application application = getApplication();
             if(application == null){
                 Log.d(Countly.TAG, "[CountlyMessagingService] getApplication() returns null: application must be non-null to init CountlyPush");
             }
             else {
-                if(mode == 0) {
-                    CountlyPush.init(application, Countly.CountlyMessagingMode.TEST);
-                } else if(mode == 1) {
-                    CountlyPush.init(application, Countly.CountlyMessagingMode.PRODUCTION);
-                }
+                CountlyConfigPush configPush = new CountlyConfigPush(application);
+                CountlyPush.init(configPush);
             }
         }